+++ src/main/java/ru/yandex/practicum/filmorate/storage/InMemoryFilmStorage.java
@@ -12,15 +12,15 @@ import java.util.Map;
 @Slf4j
 @Repository
 public class InMemoryFilmStorage implements FilmStorage {
-    private static int idCounter = 0;
-    private final Map<Integer, Film> films = new HashMap<>();
+    private static long idCounter = 0;
+    private final Map<Long, Film> films = new HashMap<>();
 
-    public boolean isFilmExist(int id) {
+    public boolean isFilmExist(long id) {
         return films.containsKey(id);
     }
 
     @Override
-    public Film getFilm(int id) {
+    public Film getFilm(long id) {
         return films.get(id);
     }
